requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
idna==3.4
joblib==1.2.0
nltk==3.8.1
numpy==1.25.0
pandas==2.0.2
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
textblob==0.17.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vaderSentiment==3.3.2
